requirements.txt
amqp==5.0.2
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.2.1
celery==5.0.5
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
imageio==2.9.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mysql-connector-python==8.0.22
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numpy==1.19.4
opencv-python==3.4.11.45
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pybind11==2.6.1
pycparser==2.20
Pygments==2.7.3
pyopencl==2020.3.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytools==2020.4.3
pytz==2020.5
PyWavelets==1.1.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
redis==3.5.3
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.12.8
tornado==6.1
traitlets==5.0.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
